# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DallyLetter Elidems - CAPS Education Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List messages (optionally by group or recipient)
      parameters:
        - name: groupId
          in: query
          schema:
            type: ["integer", "null"]
        - name: recipientId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # STUDY GROUPS
  /study-groups:
    get:
      operationId: listStudyGroups
      tags: [studyGroups]
      summary: List all study groups
      responses:
        "200":
          description: List of study groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyGroup"

    post:
      operationId: createStudyGroup
      tags: [studyGroups]
      summary: Create a study group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudyGroupBody"
      responses:
        "201":
          description: Study group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGroup"
  /study-groups/{id}:
    get:
      operationId: getStudyGroup
      tags: [studyGroups]
      summary: Get study group by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGroup"
  /study-groups/{id}/join:
    post:
      operationId: joinStudyGroup
      tags: [studyGroups]
      summary: Join a study group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGroup"

  # PAYMENTS
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments (owner can see all, student sees own)
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

    post:
      operationId: recordPayment
      tags: [payments]
      summary: Record a payment (owner/teacher)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordPaymentBody"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # NOTIFICATIONS
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create a notification (owner/teacher)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationBody"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # RAISE HAND
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, teacher, owner]
        isPrefect:
          type: boolean
        isBlocked:
          type: boolean
        phone:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        lastPaymentDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - isPrefect
        - isBlocked
        - createdAt
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderRole:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [text, voice]
        mediaUrl:
          type: ["string", "null"]
        groupId:
          type: ["integer", "null"]
        recipientId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - senderId
        - senderName
        - senderRole
        - content
        - type
        - createdAt
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        type:
          type: string
          enum: [text, voice]
        mediaUrl:
          type: ["string", "null"]
        groupId:
          type: ["integer", "null"]
        recipientId:
          type: ["integer", "null"]
      required:
        - content
        - type
    StudyGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: string
        creatorId:
          type: integer
        creatorName:
          type: string
        memberCount:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        createdAt:
          type: string
      required:
        - id
        - name
        - subject
        - creatorId
        - creatorName
        - memberCount
        - members
        - createdAt
    CreateStudyGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: string
      required:
        - name
        - subject
    Payment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        amount:
          type: number
        month:
          type: string
        year:
          type: integer
        status:
          type: string
          enum: [paid, pending, overdue]
        notes:
          type: ["string", "null"]
        recordedBy:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - studentId
        - studentName
        - amount
        - month
        - year
        - status
        - recordedBy
        - createdAt
    RecordPaymentBody:
      type: object
      properties:
        studentId:
          type: integer
        amount:
          type: number
        month:
          type: string
        year:
          type: integer
        status:
          type: string
          enum: [paid, pending, overdue]
        notes:
          type: ["string", "null"]
      required:
        - studentId
        - amount
        - month
        - year
        - status
    Notification:
      type: object
      properties:
        id:
          type: integer
        recipientId:
          type: ["integer", "null"]
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [payment_overdue, new_lesson, class_starting, system, general]
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - message
        - type
        - isRead
        - createdAt
    CreateNotificationBody:
      type: object
      properties:
        recipientId:
          type: ["integer", "null"]
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [payment_overdue, new_lesson, class_starting, system, general]
      required:
        - title
        - message
        - type
...